Externalized base URL function
